requirements.txt
langchain
langchain-community
langchain-google-genai
sentence-transformers
faiss-cpu
pypdf
streamlit
python-dotenv
tiktoken
